Do not structure a Savon request if the eventual student-list is empty ..
We build testpapers for only those students who haven't yet taken
the quiz. And therefore, those that have are filtered out. If any
remain after this filtering, then we build testpapers for them

Problem is that an empty student list makes for a "bad" web-services
request. The remote server will barf ... So, sanity checks put in place

Ideally, we should report back to the teacher that some of the students
have already taken the quiz.

+++ app/models/quiz.rb
@@ -67,9 +67,13 @@ class Quiz < ActiveRecord::Base
       end
     end # student loop 
 
-    response = (self.save) ? testpaper.compile_tex : {}
-    testpaper.destroy if response[:manifest].blank? 
-    return response
+    if testpaper.students.empty?
+      return {}
+    else
+      response = (self.save) ? testpaper.compile_tex : {}
+      testpaper.destroy if response[:manifest].blank? 
+      return response
+    end
   end 
 
   def teacher
